==> clean/02_daily_report.Rmd <==
---
title: "`r paste('Groundwater Observatory Daily Report:', Sys.Date())`"
header-includes:
   - \usepackage[table]{xcolor}
   - \usepackage{tabu}
   - \usepackage{makecell,interfaces-makecell}
output: pdf_document
---

Automagically generated at `r Sys.time()`.  

# Battery Status of Monitoring Units

```{r, echo = FALSE, message=FALSE, error=FALSE, warning=FALSE}
options(kableExtra.latex.load_packages = FALSE)
library(knitr)
library(kableExtra)
library(dplyr)
library(readr)
library(leaflet)
library(mapview) # for mapshot(): leaflet to png
library(ggplot2)
library(sp)

# load and format battery life df
battery_life_df    <- read_rds("battery_life_df.rds")
battery_life_df$id <- substr(battery_life_df$id, 2, 7)

# color coded table
battery_life_df %>%
  mutate(
    bat = cell_spec(bat, color = ifelse(bat >= 80,"#008000", ifelse(bat < 80 & bat >= 70, "#FFA500", "#ff0000") ))
  ) %>%
  kable(escape = FALSE, col.names = c("Level Sender ID", "Battery Life (\\%)", "Unit Type")) %>% 
  kable_styling(latex_options = "striped", full_width = TRUE, font_size = 20)
```



```{r, out.width = "100%", echo = FALSE, message=FALSE, error=FALSE, warning=FALSE}
# get the coordinates of monitoring well stations and extract the necessary info
cs_coords <- read_tsv("https://raw.githubusercontent.com/richpauloo/cosumnes_shiny/master/clean/dependencies/elev.txt")
cs_coords <- cs_coords[ , c("ls_id", "lat", "lng")] %>% 
  filter(!is.na(ls_id)) %>% 
  rename(id = ls_id) %>% 
  mutate(id = as.character(id))

# units == Sender
df <- left_join(battery_life_df, cs_coords, by = "id") %>% 
  filter(!is.na(lat) & unit == "Sender")

# get components of SPDF
dat    <- df[, 1:3]
coords <- df[, c("lng", "lat")] 

# make into SPDF
spdf <- SpatialPointsDataFrame(coords = coords, data = dat, proj4string = CRS("+init=epsg:4326 +proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0"))


# bins for the battery life 
pal <- colorBin(palette = c('#d7191c','#fdae61','#ffffbf','#abd9e9','#2c7bb6'), # red to green
                domain = spdf$bat, bins = c(0,70,80,90,100))

# mape the leaflet
bat_map <- spdf %>% 
  leaflet() %>% 
  addProviderTiles(providers$CartoDB) %>% 
  addCircleMarkers(popup = spdf$bat,
                   fillColor = ~pal(bat), 
                   fillOpacity = 1,
                   color = "black") %>% 
  addLegend(pal = pal, values = ~bat, opacity = 1, title = "Battery Life (%)") %>% 
  addLabelOnlyMarkers(label = ~id, 
                      labelOptions = labelOptions(noHide = TRUE,
                                                  textOnly = TRUE,
                                                  direction = "top",
                                                  textsize = 30))
  

# save to png
mapshot(bat_map, file = "ls_bat_map.png")



############################################



# units == Logger
df <- left_join(battery_life_df, cs_coords, by = "id") %>% 
  filter(!is.na(lat) & unit == "Logger")

# get components of SPDF
dat    <- df[, 1:3]
coords <- df[, c("lng", "lat")] 

# make into SPDF
spdf <- SpatialPointsDataFrame(coords = coords, data = dat, proj4string = CRS("+init=epsg:4326 +proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0"))


# bins for the battery life 
pal <- colorBin(palette = c('#d7191c','#fdae61','#ffffbf','#abd9e9','#2c7bb6'), # red to green
                domain = spdf$bat, bins = c(0,70,80,90,100))

# mape the leaflet
bat_map <- spdf %>% 
  leaflet() %>% 
  addProviderTiles(providers$CartoDB) %>% 
  addCircleMarkers(popup = spdf$bat,
                   fillColor = ~pal(bat), 
                   fillOpacity = 1,
                   color = "black") %>% 
  addLegend(pal = pal, values = ~bat, opacity = 1, title = "Battery Life (%)") %>% 
  addLabelOnlyMarkers(label = ~id, 
                      labelOptions = labelOptions(noHide = TRUE,
                                                  textOnly = TRUE,
                                                  direction = "top",
                                                  textsize = 30))

mapshot(bat_map, file = "ll_bat_map.png")

############################################


# units == Baro
df <- left_join(battery_life_df, cs_coords, by = "id") %>% 
  filter(!is.na(lat) & unit == "Baro")

# get components of SPDF
dat    <- df[, 1:3]
coords <- df[, c("lng", "lat")] 

# make into SPDF
spdf <- SpatialPointsDataFrame(coords = coords, data = dat, proj4string = CRS("+init=epsg:4326 +proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0"))


# bins for the battery life 
pal <- colorBin(palette = c('#d7191c','#fdae61','#ffffbf','#abd9e9','#2c7bb6'), # red to green
                domain = spdf$bat, bins = c(0,70,80,90,100))

# mape the leaflet
bat_map <- spdf %>% 
  leaflet() %>% 
  addProviderTiles(providers$CartoDB) %>% 
  addCircleMarkers(popup = spdf$bat,
                   fillColor = ~pal(bat), 
                   fillOpacity = 1,
                   color = "black") %>% 
  addLegend(pal = pal, values = ~bat, opacity = 1, title = "Battery Life (%)") %>% 
  addLabelOnlyMarkers(label = ~id, 
                      labelOptions = labelOptions(noHide = TRUE,
                                                  textOnly = TRUE,
                                                  direction = "top",
                                                  textsize = 30))

mapshot(bat_map, file = "bl_bat_map.png")
```

# Level Senders
```{r, out.width = "100%", echo = FALSE, message=FALSE, error=FALSE, warning=FALSE}
include_graphics("ls_bat_map.png")
```

# Level Loggers
```{r, out.width = "100%", echo = FALSE, message=FALSE, error=FALSE, warning=FALSE}
include_graphics("ll_bat_map.png")
```

# Baro Logger
```{r, out.width = "100%", echo = FALSE, message=FALSE, error=FALSE, warning=FALSE}
include_graphics("bl_bat_map.png")
```


***  


# Last 30 days of Received Data
```{r, out.width="100%", echo = FALSE, message=FALSE, error=FALSE, warning=FALSE}
# ggplot of last 30 days of data
ggp <- read_rds("ggp.rds")
ggp

# text_df <- ggp$data %>% 
#   filter(!is.na(level)) %>% 
#   group_by(id) %>% 
#   mutate(pos = 1:n()) %>% 
#   ungroup() %>% 
#   dplyr::arrange(desc(dt)) %>% 
#   filter(pos == 1)
# 
# ggp$data %>% 
#   ggplot() + 
#   geom_line(aes(dt, level, color = id)) + 
#   geom_label_repel(data = text_df, aes(dt, level, label = id, color = id), hjust = 1, nudge_y = 1)
```
