==> clean/03_run.R <==
# set the file path of the directory where cleaning scripts live, and source the starting file
source("C:/Users/rpauloo/Documents/GitHub/cosumnes_shiny/clean/01_query_sqlite_clean_and_push_to_cloud_SQL.R")
